pyproject.toml
[tool.poetry]
name = "fastapi-git-p8"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_git_p8"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
